% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_new_gp.R
\name{calc_new_gp}
\alias{calc_new_gp}
\title{Describes the Growth of Chickpea}
\usage{
calc_new_gp(current_growing_points, gp_rr, max_gp, mean_air_temp)
}
\arguments{
\item{current_growing_points}{Are the current number of growing points at an iteration period}

\item{gp_rr}{Is the rate of increase of chickpea growing points in an iteration period of one day}

\item{max_gp}{Are the maximum number of chickpea growing points for each cultivar}

\item{mean_air_temp}{Is average daily temperature in Celcius}
}
\value{
Integer giving the number of new growing points formed in an iteration period of one day
}
\description{
'calc_new_gp()' describes the growth of chickpea in terms of development of growing
points. Upon germination, each germinated seed produces one growing point that multiplies at a rate
that is a function of temperature and is limited by proximity to a maximum growing point density.
}
