% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_weather.R
\name{format_weather}
\alias{format_weather}
\title{Format weather data into a ascotraceR.weather object for use in the spore
dispersal models}
\usage{
format_weather(
  x,
  YYYY = NULL,
  MM = NULL,
  DD = NULL,
  hh = NULL,
  mm = NULL,
  POSIXct_time = NULL,
  time_zone = NULL,
  temp,
  rain,
  ws,
  wd,
  wd_sd,
  station,
  lon = NULL,
  lat = NULL,
  r = NULL,
  lonlat_file = NULL
)
}
\arguments{
\item{x}{a \code{\link{data.frame}} object of weather station data for
formatting.}

\item{YYYY}{column name or index in \code{x} that refers to the year when the
weather was logged.}

\item{MM}{column name or index in \code{x} that refers to the month (numerical)
when the weather was logged.}

\item{DD}{column name or index in \code{x} that refers to the day of month when
the weather was logged.}

\item{hh}{column name or index in \code{x} that refers to the hour (24 hour)
when the weather was logged.}

\item{mm}{column name or index in \code{x} that refers to the minute when the
weather was logged.}

\item{POSIXct_time}{column name or index in \code{x} which contains a \code{POSIXct}
formatted time. This can be used instead of arguments \code{YYYY}, \code{MM}, \code{DD},
\code{hh}, \code{mm.}}

\item{time_zone}{time zone (Olsen time zone format) where the weather station
is located. May be in a column or supplied as a character string.
Optional, see also \code{r}. See details.}

\item{temp}{column name or index in \code{x} that refers to temperature in degrees
Celsius}

\item{rain}{column name or index in \code{x} that refers to rainfall in mm.}

\item{ws}{column name or index in \code{x} that refers to wind speed in km / h.}

\item{wd}{column name or index in \code{x} that refers to wind direction in degrees.}

\item{wd_sd}{column name or index in \code{x} that refers to wind speed in km / h
character.  This is only applicable if weather data is already summarised to
hourly increments. See details.}

\item{station}{column name or index in \code{x} that refers to the weather station
name or identifier. See details.}

\item{lon}{column name or index in \code{x} that refers to weather station's
longitude.  See details.}

\item{lat}{column name or index in \code{x} that refers to weather station's
latitude.  See details.}

\item{r}{Spatial raster which is intended to be used with this weather data
for use in the blackspot model. Used to fetch time_zone if it is not
supplied in data. Optional, see also \code{time_zone}.}

\item{lonlat_file}{a file path to a csv which included station name/id and
longitude and latitude coordinates if they are not supplied in data.
Optional, see also \code{lon} and \code{lat}.}
}
\value{
A \code{ascotraceR.weather} object (an extension of
\CRANpkg{data.table}) containing the supplied weather aggregated to each
hour in a suitable format for use with \code{\link{trace_asco}}
containing the following columns:
\tabular{rl}{
\strong{times}: \tab Time in POSIXct format \cr
\strong{rain}: \tab Rainfall in mm \cr
\strong{ws}: \tab Wind speed in km / h \cr
\strong{wd}: \tab Wind direction in compass degrees \cr
\strong{wd_sd}: \tab Wind direction standard deviation in compass degrees \cr
\strong{lon}: \tab Station longitude in decimal degrees \cr
\strong{lat}: \tab Station latitude in decimal degrees \cr
\strong{station}: \tab Unique station identifying name \cr
\strong{YYYY}: \tab Year \cr
\strong{MM}: \tab Month \cr
\strong{DD}: \tab Day \cr
\strong{hh}: \tab Hour \cr
\strong{mm}: \tab Minute \cr
}
}
\description{
Formats raw weather data into an object suitable for use in
\code{\link{trace_asco}} function ensuring that the supplied weather data meet
the requirements of the model to run.
Internal support for multithreaded operations is provided through
\CRANpkg{future}.  If more than one station is present, the process
can be made faster by using \code{\link[future]{plan}}.
}
\details{
\code{time_zone}
All weather stations must fall within the same time zone.  If the required
stations are located in differing time zones, separate \code{ascotraceR.weather}
objects must be created for each time zone.  If a raster object of
previous crops is provided that spans time zones, an error will be emitted.

\code{wd_sd}
If weather data is provided in hourly increments, a column with the standard
deviation of the wind direction over the hour is required to be provided.
If the weather data are sub-hourly, these data will be calculated and
returned automatically.

\code{lon}, \code{lat} and \code{lonlat_file}
If \code{x} provides longitude and latitude values for station locations, these
may be specified in the \code{lon} and \code{lat} columns.  If gps coordinates are
not relevant to the study location \code{NA} can be specified and the function
will drop these column variables.  If these data are not included, (\code{NULL})
a separate file may be provided that contains the longitude, latitude and
matching station name to provide station locations in the final
\code{ascotraceR.weather} object that is created by specifying the file path to
a \acronym{CSV file} using \code{lonlat_file}.
}
\examples{
# Fake weather data files for testing and examples have been included in
# \pkg{ascotraceR}.  The weather data files both are of the same format, so
# they will be combined for formatting here.

Billa_Billa_weather <-
  read_csv(system.file("extdata",
            "2020_Billa_Billa_weather.csv", package = "ascotraceR")) \%>\%
  mutate(
    wind_direction = case_when(
      max_wind_direction == "N" ~ "0",
      max_wind_direction == "NbE" ~ "11.25",
      max_wind_direction == "NNE" ~ "22.5",
      max_wind_direction == "NEbN" ~ "33.75",
      max_wind_direction == "NE" ~ "45",
      max_wind_direction == "NEbE" ~ "56.25",
      max_wind_direction == "ENE" ~ "67.5",
      max_wind_direction == "EbN" ~ "73.5",
      max_wind_direction == "E" ~ "90",
      max_wind_direction == "EbS" ~ "101.2",
      max_wind_direction == "ESE" ~ "112.5",
      max_wind_direction == "SEbE" ~ "123.8",
      max_wind_direction == "SE" ~ "135.1",
      max_wind_direction == "SEbS" ~ "146.3",
      max_wind_direction == "SSE" ~ "157.6",
      max_wind_direction == "SbE" ~ "168.8",
      max_wind_direction == "S" ~ "180",
      max_wind_direction == "SbW" ~ "191.2",
      max_wind_direction == "SSW" ~ "202.5",
      max_wind_direction == "SWbS" ~ "213.8",
      max_wind_direction == "SW" ~ "225",
      max_wind_direction == "SWbW" ~ "236.2",
      max_wind_direction == "WSW" ~ "247.5",
      max_wind_direction == "WbS" ~ "258.8",
      max_wind_direction == "W" ~ "270",
      max_wind_direction == "WbN" ~ "281.2",
      max_wind_direction == "WNW" ~ "292.5",
      max_wind_direction == "NWbW" ~ "303.8",
      max_wind_direction == "NW" ~ "315",
      max_wind_direction == "NWbN" ~ "326.2",
      max_wind_direction == "NNW" ~ "337.5",
      max_wind_direction == "NbW" ~ "348.8",
      TRUE ~ max_wind_direction
    )
  ) \%>\%
  mutate(wind_direction = as.numeric(wind_direction))
Billa_Billa_station_data <-
  system.file("extdata", "Billa_Billa_stat_dat.csv", package = "ascotraceR")

weather <- format_weather(
   x = Billa_Billa_weather,
   POSIXct_time = "local_time",
   ws = " avg_wind_speed (m/s)",
   wd_sd = "wd_sd",
   rain = "rain",
   wd = "wd",
   temp = "mean_daily_temperature"
   station = "location",
   time_zone = "Australia/Brisbane",
   lonlat_file = Billa_Billa_station_data
)

}
