% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercept_spores.R
\name{intercept_spores}
\alias{intercept_spores}
\title{Calculate spore interception parameter}
\usage{
intercept_spores(
  spore_interception_parameter = 0.00006,
  max_growing_points_limit,
  max_new_growing_points_limit
)
}
\arguments{
\item{spore_interception_parameter}{multiplier to determine spore interception parameter.
Defaults to \code{0.00006}}

\item{max_growing_points_limit}{Are the maximum number of growing points per square metre.
Defaults to \code{15000}}

\item{max_new_growing_points_limit}{A \emph{numeric} number indicating number of new growing points
developed per square metre in an iteration period of one day.  Defaults to \code{350}
@keywords internal
@noRd}
}
\description{
The parameter which controls the relationship between uninfected growing points per area
and probability of a spore causing an infection.
The spore_interception_parameter gives an estimate of spores not producing
infections
}
