% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_some_infective.R
\name{make_some_infective}
\alias{make_some_infective}
\title{Make infected growing points sources for spore dispersal}
\usage{
make_some_infective(daily_vals, latent_period = 200)
}
\arguments{
\item{daily_vals}{list of the current day's values and paddock data.table}

\item{spore_packet}{data.table with three variables, 'x', 'y' and
'spores_per_packet'}
}
\value{
updated daily_vals list
}
\description{
Make infected growing points sources for spore dispersal
}
\examples{
new_daily_vals <-
   make_some_infective(
      spore_packet = newly_infective,
      daily_vals = daily_vals_list)
}
