% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/max_growing_points.R
\name{max_growing_points}
\alias{max_growing_points}
\title{Maximum number of growing points chickpea plant achieved}
\usage{
max_growing_points()
}
\arguments{
\item{per_area}{is given area (observatoin quadrat or paddock) in metres}

\item{max_growing_points_limit}{is the limit on number of growing points chickpea
will ever achieve. It depends on chickpa cultivar}

\item{min_growing_points}{are the minimum number of growing points}
}
\value{
a \emph{numeric} vector
}
\description{
'max_growing_points()' calculates maximum number of growing points chickpea plant
achived. The growth of chickpea is described in terms of number of growing points
}
