% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_growing_points.R
\name{new_growing_points}
\alias{new_growing_points}
\title{Describes the Growth of Chickpea}
\usage{
new_growing_points(current_growing_points, mean_air_temp)
}
\arguments{
\item{current_growing_points}{Are the current number of growing points at an iteration period}

\item{mean_air_temp}{Is average daily temperature in Celcius}
}
\value{
numeric vector giving the number of new growing points formed that day
}
\description{
'new_growing_points()' describes the growth of chickpea in terms of development of growing
points. Upon germination, each germinated seed produces one growing point that multiplies at a rate
that is a function of temperature and is limited by proximity to a maximum growing point density.
}
