% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_day.R
\name{one_day}
\alias{one_day}
\title{Simulates ascochyta spore dispersal for a single day increment}
\usage{
one_day(i_date, daily_vals, weather_dat)
}
\arguments{
\item{weather_dat}{}
}
\value{

}
\description{
Simulates ascochyta spore dispersal for a single day increment
}
