% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/potentially_effective_spores.R
\name{potentially_effective_spores}
\alias{potentially_effective_spores}
\title{Estimate the potentially effective spore}
\usage{
potentially_effective_spores(
  sporesPerInfectiveGPPerWetHour,
  max_interception_probability,
  paddock_infected_gp
)
}
\arguments{
\item{sporesPerInfectiveGPPerWetHour}{estimated number of spores produced in one hour from a single chickpea growing point}

\item{max_interception_probability}{maximum interception probability}

\item{paddock_infected_gp}{integer value giving paddockInfectiveGrowingPoints at paddock coordinates}

\item{infected_source_address}{vector of length two giving the paddock coordinates with ascochyta infection}
}
\value{
a double providing the potentially infective spores at the paddock_infected_gp coordinates
}
\description{
Estimate the potentially effective spore
}
