% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracer_plot.R
\name{tracer_plot}
\alias{tracer_plot}
\title{plot one day from trace_asco output as a tile plot}
\usage{
tracer_plot(dat, day, tiles = "sporulating_gp")
}
\arguments{
\item{dat}{nested list of ascotraceR class}

\item{day}{integer, day of the simulation to plot}

\item{tiles}{what to response for the plot to render, options: sporulating_gp, new_gp,noninfected_gp}
}
\value{
ggplot
}
\description{
plot one day from trace_asco output as a tile plot
}
\examples{
test5 <- trace_asco(
  weather = Ascotracer::newM_weather,
  paddock_length = 100,
  paddock_width = 100,
  initial_infection = "1998-06-10",
  sowing_date = as.POSIXct("1998-06-09"),
  harvest_date = as.POSIXct("1998-06-09") + lubridate::ddays(100),
  time_zone = "Australia/Perth",
  primary_infection_foci = "center"

  tracer_plot(test5,102))
}
