% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_paddock_gp.R
\name{update_paddock_gp}
\alias{update_paddock_gp}
\title{Update chickpea growing points over a whole paddock}
\usage{
update_paddock_gp(paddock_coords, mean_air_temp, gp_rr, max_gp)
}
\arguments{
\item{paddock_coords}{a data.frame of paddock coordinates containing a column variable "noninfected_gp}

\item{mean_air_temp}{mean daily air temperature}

\item{gp_rr}{growing points replication rate}

\item{max_gp}{maximum growing points}
}
\value{
a vector of updated growing points at each paddock location
}
\description{
Update chickpea growing points over a whole paddock
}
