% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/which_paddock_row.R
\name{which_paddock_row}
\alias{which_paddock_row}
\title{Returns the row numbers of an x y data frame}
\usage{
which_paddock_row(paddock, query)
}
\arguments{
\item{paddock}{The template data frame you wish to know the row numbers of}

\item{query}{data.frame with column names 'x' and 'y' for which you want to know
the row number of in the paddock data.table}
}
\value{
vector of row numbers
}
\description{
Returns the row numbers of an x y data frame
}
\examples{
pdk <- as.data.table(expand.grid(x = 1:100,
                                 y = 1:100))
qry <- pdk[sample(1:nrow(pdk), 5),]
which_paddock_row(pdk, qry)
}
