% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_distance.R
\name{wind_distance}
\alias{wind_distance}
\title{Wind distance dispersal parameter}
\usage{
wind_distance(
  average_wind_speed_in_fifteen_minutes,
  wind_cauchy_multiplier = 0.1,
  PSPH = 1
)
}
\arguments{
\item{average_wind_speed_in_fifteen_minutes}{A numeric value in m/s}

\item{wind_cauchy_multiplier}{A Cauchy scaling parameter}

\item{PSPH}{A numeric vector estimated from '.estimate_spore_discharge()'}

\item{return}{Distance conidia dispersed by wind from the source of infection}
}
\description{
\code 'wind_distance()' determines distance conidia disperse, in metres, by wind
Conidia are assumed to spread from the centre of each subunit.
The destination subunit, where conidia land, could be the same subunit or another subunit within
or outside the paddock.
If conidia land outside the paddock, the model can simulate the spread outside the paddock
when the wrapping option is chosen.
When the wrapping option is not chosen, the model simulate disease spread within the paddock
}
\examples{
 wind_distance(10, PSPH = 10) # output is 10 estimates
wind_distance(10) # output is a single estimate
}
