% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascotraceR-package.R
\docType{package}
\name{ascotraceR-package}
\alias{ascotraceR}
\alias{ascotraceR-package}
\title{ascotraceR: A Spatiotemporal Model to Simulate the Spread of Ascochyta Blight in Chickpea}
\description{
A spatiotemporal model that simulates the spread of Ascochyta blight in chickpea fields based on location-specific weather conditions.
}
\author{
\strong{Maintainer}: Paul Melloy \email{paul.melloy@usq.edu.au} (\href{https://orcid.org/0000-0003-4253-7167}{ORCID})

Authors:
\itemize{
  \item Ihsanul Khaliq \email{ihsan.khaliq@usq.edu.au} (\href{https://orcid.org/0000-0003-4171-0917}{ORCID})
  \item Adam H. Sparks \email{adamhsparks@gmail.com} (\href{https://orcid.org/0000-0002-0061-8359}{ORCID}) [conceptor]
}

Other contributors:
\itemize{
  \item Grains Research and Development Corporation (GRDC) Project USQ1903-003RTX [funder, copyright holder]
  \item University of Southern Queensland [copyright holder]
  \item Western Australia Agriculture Authority (WAAA) [copyright holder]
}

}
\keyword{internal}
