% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spore_interception_parameter.R
\name{spore_interception_parameter}
\alias{spore_interception_parameter}
\title{The parameter which controls the relationship between uninfected growing points per area
and probability of a spore causing an infection}
\usage{
spore_interception_parameter()
}
\arguments{
\item{max_growing_points_limit}{Are the maximum number of growing points chickpea can achieve}

\item{max_new_growing_points_limit}{A \emph{numeric} number giving number of new growing points
@keywords internal
@noRd}
}
\description{
'spore_interception_parameter()' gives an estimate of spores not spores not producing
infections
}
