% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_trace.R
\name{summarise_trace}
\alias{summarise_trace}
\title{Summarise a trace_asco output nested list}
\usage{
summarise_trace(trace)
}
\arguments{
\item{trace}{a nested list output from \code{\link[=trace_asco]{trace_asco()}}}
}
\value{
A \CRANpkg{data.table} summarising the model's output for a paddock
on a daily time-step with the area under the disease progress curve
(\acronym{AUDPC}) at the paddock level for the simulation's run.
}
\description{
Creates a paddock-level summary \CRANpkg{data.table} from the output of
\code{\link[=trace_asco]{trace_asco()}} on a daily time-step.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
Newmarracarra <-
   read.csv(system.file("extdata",
            "1998_Newmarracarra_weather_table.csv", package = "ascotraceR"))
station_data <-
   system.file("extdata", "stat_dat.csv", package = "ascotraceR")

weather_dat <- format_weather(
   x = Newmarracarra,
   POSIXct_time = "Local.Time",
   temp = "mean_daily_temp",
   ws = "ws",
   wd_sd = "wd_sd",
   rain = "rain_mm",
   wd = "wd",
   station = "Location",
   time_zone = "Australia/Perth",
   lonlat_file = station_data)

traced <- trace_asco(
  weather = weather_dat,
  paddock_length = 100,
  paddock_width = 100,
  initial_infection = "1998-06-10",
  sowing_date = as.POSIXct("1998-06-09"),
  harvest_date = as.POSIXct("1998-06-09") + lubridate::ddays(100),
  time_zone = "Australia/Perth",
  primary_infection_foci = "centre")

tidied <- summarise_trace(traced)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=trace_asco]{trace_asco()}}, \code{\link[=tidy_trace]{tidy_trace()}}
}
