% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_asco.R
\name{trace_asco}
\alias{trace_asco}
\title{Estimate the spread of ascochyta blight on chickpea}
\usage{
trace_asco(weather, paddock_length, paddock_width, sowing_date)
}
\arguments{
\item{weather}{weather data to inform the disease dynamics and crop
maturity through the chickpea growing season.}

\item{paddock_length}{length of paddock in meters (x)}

\item{paddock_width}{width of paddock in meters (y)}

\item{sowing_date}{sowing date of chickpea seed and the start of the
Ascochyta tracer model. Assumes there is sufficient soil moisture to
induce germination}
}
\value{
a x y \code{data.frame} providing the paddock coordinates and estimated
severity of ascochyta at the respectic location
}
\description{
\code{trace_asco} estimates the spatial spread through a chickpea crop
during the growing season.
}
\examples{
ta1 <- trace_asco(
  weather = weather_dat,
  paddock_length = 100,
  paddock_width = 100,
  sowing_date = "1998-03-09"
  )
}
