% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_asco.R
\name{trace_asco}
\alias{trace_asco}
\title{Simulates the spread of Ascochyta blight in a chickpea field}
\usage{
trace_asco(
  weather,
  paddock_length,
  paddock_width,
  sowing_date,
  harvest_date,
  initial_infection,
  seeding_rate = 40,
  gp_rr = 0.0065,
  max_gp_lim = 15000,
  max_new_gp = 350,
  latent_period_cdd = 200,
  time_zone = "UTC",
  primary_infection_foci = "random",
  primary_infection_intensity = 1,
  n_foci = 1,
  spores_per_gp_per_wet_hour = 0.22
)
}
\arguments{
\item{weather}{weather data, recorded by a local weather station, over a
chickpea growing season for the model operation.}

\item{paddock_length}{length of a paddock in metres (y).}

\item{paddock_width}{width of a paddock in metres (x).}

\item{sowing_date}{a character string of a date value indicating sowing date
of chickpea seed and the start of the ascotraceR model. Preferably in
ISO8601 format (YYYY-MM-DD), \emph{e.g.} \dQuote{2020-04-26}. Assumes there
is sufficient soil moisture to induce germination and start the crop
growing season.}

\item{harvest_date}{a character string of a date value indicating harvest
date of chickpea crop, which is also the last day to run the ascotraceR
model. Preferably in ISO8601 format (YYYY-MM-DD), \emph{e.g.},
\dQuote{2020-04-26}.}

\item{initial_infection}{refers to initial or primary infection on seedlings,
resulting in the production of infected growing points.}

\item{seeding_rate}{indicate the rate at which chickpea seed is sown per
square metre. Defaults to \code{40}.}

\item{gp_rr}{refers to rate of increase in chickpea growing points per degree
Celsius per day. Defaults to \code{0.0065}.}

\item{max_gp_lim}{Maximum number of chickpea growing points (meristems)
allowed per square metre. Defaults to \code{15000}.}

\item{max_new_gp}{Maximum number of new chickpea growing points (meristems),
which develop per day, per square metre. Defaults to \code{350}.}

\item{latent_period_cdd}{latent period in cumulative degree days (sum of
daily temperature means) is the period between infection and production of
lesions on susceptible growing points. Defaults to \code{200}.}

\item{time_zone}{refers to time in Coordinated Universal Time (UTC).}

\item{primary_infection_foci}{refers to the inoculated coordinates where the
epidemic starts. Accepted inputs are: \code{centre} or \code{random} (Default) or a
data.frame with column names 'x', 'y' and 'load'. The \code{data.frame} inputs
inform the model of specific grid cell/s coordinates where the epidemic
should begin. The 'load' column is optional and can specify the
\code{primary_infection_intensity} for each coordinate.}

\item{primary_infection_intensity}{The intensity of the starting epidemic as
described by the number of number of sporulating growing points.}

\item{n_foci}{only relevant when \code{primary_infection_foci = "random"} and
notes the number of \code{primary_infection_foci} at initial infection.}

\item{spores_per_gp_per_wet_hour}{Number of spores produced per sporulating
growing point each wet hour. Also known as the \code{spore_rate}. Value is
dependent on the susceptibility of the host genotype.}
}
\value{
\link{list} object with two elements. The first is a \link{matrix} of values
calculated by the model where each row is a unique observation that contains:
\describe{
\item{day}{numeric day of the simulation run}
\item{x}{location of metre squared area on x-axis in simulated paddock}
\item{y}{location of metre squared area on y-axis in simulated paddock}
\item{new_gp}{new growing points on \code{day} for location \code{x} and \code{y}}
\item{susceptible_gp}{susceptible growing points on \code{day} for location \code{x} and \code{y}}
\item{exposed_gp}{exposed growing points on \code{day} for location \code{x} and \code{y}}
\item{infectious_gp}{infectious points on \code{day} for location \code{x} and \code{y}}
\item{infected_coords}{coordinates for infected quadrats on \code{day}}
\item{exposed_gps}{exposed growing points in the latent period phase of infection on \code{day} for location \code{x} and \code{y}}
\item{spores_per_packet}{}
\item{cdd_at_infection}{cumulative degree days at infection}
}
The second is a \code{data.table} containing:
\describe{
\item{i_date}{}
\item{i_day}{}
\item{i_date}{}
\item{cwh}{cumulative wet hours on \code{day}}
\item{cr}{cumulative rainfall on \code{day}}
\item{cdd}{cumulative degree days on \code{day}}
}
}
\description{
\code{trace_asco} simulates the spatiotemporal development of Ascochyta blight in
a chickpea field over a growing season.
}
\examples{

# First weather data needs to be imported and formatted with `format_weather`
Newmarracarra <-
   read.csv(system.file("extdata",
            "1998_Newmarracarra_weather_table.csv", package = "ascotraceR"))
station_data <-
   system.file("extdata", "stat_dat.csv", package = "ascotraceR")

weather_dat <- format_weather(
   x = Newmarracarra,
   POSIXct_time = "Local.Time",
   temp = "mean_daily_temp",
   ws = "ws",
   wd_sd = "wd_sd",
   rain = "rain_mm",
   wd = "wd",
   station = "Location",
   time_zone = "Australia/Perth",
   lonlat_file = station_data)

# Now the `trace_asco` function can be run to simulate disease spread
traced <- trace_asco(
  weather = weather_dat,
  paddock_length = 100,
  paddock_width = 100,
  initial_infection = "1998-06-10",
  sowing_date = as.POSIXct("1998-06-09"),
  harvest_date = as.POSIXct("1998-06-09") + lubridate::ddays(70), # run the model for 70 days
  time_zone = "Australia/Perth",
  gp_rr = 0.0065,
  primary_infection_intensity = 1000,
  spores_per_gp_per_wet_hour = 0.22,
  primary_infection_foci = "centre")

traced[[70]] # extracts the model output for 70 _days_
}
