% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_asco.R
\name{trace_asco}
\alias{trace_asco}
\title{Simulates the spread of Ascochyta blight in a chickpea field}
\usage{
trace_asco(
  weather,
  paddock_length,
  paddock_width,
  sowing_date,
  harvest_date,
  initial_infection,
  seeding_rate = 40,
  gp_rr = 0.0065,
  max_gp_lim = 15000,
  max_new_gp = 350,
  latent_period_cdd = 200,
  time_zone = "UTC",
  primary_infection_foci = "random",
  primary_infection_intensity = 1,
  n_foci = 1,
  spores_per_gp_per_wet_hour = 0.22
)
}
\arguments{
\item{weather}{weather data, recorded by a local weather station, over a chickpea
growing season for the model operation}

\item{paddock_length}{length of a paddock in metres (y)}

\item{paddock_width}{width of a paddock in metres (x)}

\item{sowing_date}{a character string of a date value indicating sowing
date of chickpea seed and the start of the ascotraceR model. Preferably
in \acronym{ISO8601} format (YYYY-MM-DD), \emph{e.g.} \dQuote{2020-04-26}.
Assumes there is sufficient soil moisture to induce germination and start the
crop growing season.}

\item{harvest_date}{a character string of a date value indicating harvest date of
chickpea crop, which is also the last day to run the ascotraceR model. Preferably in
\acronym{ISO8601} format (YYYY-MM-DD), \emph{e.g.} \dQuote{2020-04-26}.}

\item{seeding_rate}{indicate the rate at which chickpea seed is sown per
square metre. Defaults to \code{40}}

\item{gp_rr}{refers to rate of increase in chickpea growing points
per degree Celsius per day. Defaults to \code{0.0065}}

\item{max_gp_lim}{Maximum number of chickpea growing points (meristems) allowed
per square meter. Defaults to \code{15000}.}

\item{max_new_gp}{Maximum number of new chickpea growing points (meristems)
which develop per day, per square meter. Defaults to \code{350}.}

\item{latent_period_cdd}{latent period in cumulative degree days (sum of
daily temperature means) is the period between infection and production of
lesions on susceptible growing points. Defaults to \code{200}
@param initial_infection refers to initial or primary infection on seedlings,
resulting in the production of infected growing points
@param time_zone refers to time in Coordinated Universal Time (UTC)
@param spores_per_gp_per_wet_hour Number of spores produced per sporulating growing point each wet hour.
Also known as the 'spore_rate'. Value is dependent on the susceptibility of the host genotype.}

\item{primary_infection_foci}{it refers to the inoculated quadrat
located at the centre of the paddock from where disease spreads
Defaults to \code{"centre"}}

\item{primary_infection_intensity}{The intensity of the starting epidemic as
described by the number of number of sporulating growing points.}
}
\value{
a x y \code{data.frame} simulating the spread of Ascochyta blight in a
chickpea paddock
}
\description{
\code{trace_asco} simulates the spatiotemporal development of Ascochyta blight in a chickpea field
over a growing season
}
\examples{
ta1 <- trace_asco(
  weather = weather_dat,
  paddock_length = 100,
  paddock_width = 100,
  sowing_date = "1998-03-09"
  )
}
