% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_asco.R
\name{trace_asco}
\alias{trace_asco}
\title{Simulates the spread of Ascochyta blight in a chickpea field}
\usage{
trace_asco(
  weather,
  paddock_length,
  paddock_width,
  sowing_date,
  harvest_date,
  initial_infection,
  seeding_rate = 40,
  gp_rr = 0.0065,
  max_gp_lim = 5000,
  max_new_gp = 350,
  latent_period_cdd = 150,
  time_zone = "UTC",
  primary_infection_foci = "random",
  primary_inoculum_intensity = 1,
  n_foci = 1,
  spores_per_gp_per_wet_hour = 0.22
)
}
\arguments{
\item{weather}{weather data for a representative chickpea paddock that
contains weather data for a complete chickpea growing season for the
model's operation.}

\item{paddock_length}{length of a paddock in metres (y).}

\item{paddock_width}{width of a paddock in metres (x).}

\item{sowing_date}{a character string of a date value indicating sowing date
of chickpea seed and the start of the \sQuote{ascotraceR} model. Preferably
in ISO8601 format (YYYY-MM-DD), \emph{e.g.} \dQuote{2020-04-26}. Assumes
there is sufficient soil moisture to induce germination and start the crop
growing season.}

\item{harvest_date}{a character string of a date value indicating harvest
date of chickpea crop, which is also the last day to run the
\sQuote{ascotraceR} model. Preferably in ISO8601 format (YYYY-MM-DD),
\emph{e.g.}, \dQuote{2020-04-26}.}

\item{initial_infection}{a character string of a date value referring to the
initial or primary infection on seedlings, resulting in the production of
infected growing points.}

\item{seeding_rate}{indicate the rate at which chickpea seed is sown per
square metre. Defaults to \code{40}.}

\item{gp_rr}{refers to rate of increase in chickpea growing points per degree
Celsius per day. Defaults to \code{0.0065}.}

\item{max_gp_lim}{Maximum number of chickpea growing points (meristems)
allowed per square metre. Defaults to \code{5000}.}

\item{max_new_gp}{Maximum number of new chickpea growing points (meristems),
which develop per day, per square metre. Defaults to \code{350}.}

\item{latent_period_cdd}{latent period in cumulative degree days (sum of
daily temperature means) is the period between infection and production of
lesions on susceptible growing points. Defaults to \code{150}.}

\item{time_zone}{refers to time in Coordinated Universal Time (UTC).}

\item{primary_infection_foci}{refers to the inoculated coordinates where the
epidemic starts. Accepted inputs are: \code{centre} or \code{random} (Default) or a
data.frame with column names 'x', 'y' and 'load'. The \code{data.frame} inputs
inform the model of specific grid cell/s coordinates where the epidemic
should begin. The 'load' column is optional and can specify the
\code{primary_inoculum_intensity} for each coordinate.}

\item{primary_inoculum_intensity}{The intensity of the starting epidemic is
described by the number of sporulating growing points (from seed infection
and/or volunteers) and/or sporulating lesions from crop debris at
\code{initial_infection}.}

\item{n_foci}{only relevant when \code{primary_infection_foci = "random"} and
notes the number of \code{primary_infection_foci} at initial infection.}

\item{spores_per_gp_per_wet_hour}{Number of spores produced per sporulating
growing point each wet hour. Also known as the \code{spore_rate}. Value is
dependent on the susceptibility of the host genotype.}
}
\value{
A nested \code{list} object where each sub-list contains daily data
for the day \code{i_day} (the model's iteration day) generated by the model
including:
\itemize{
\item \strong{paddock}, an 'x' 'y' \CRANpkg{data.table} containing:
\itemize{
\item \strong{x}, location of quadrat on x-axis in paddock,
\item \strong{y}, location of quadrat on y-axis in paddock,
\item \strong{new_gp}, new growing points produced in the last 24 hours,
\item \strong{susceptible_gp}, susceptible growing points in the last 24 hours,
\item \strong{exposed_gp}, exposed growing points in the last 24 hours,
\item \strong{infectious_gp}, infectious growing points in the last 24 hours,
}
\item \strong{i_day}, model iteration day,
\item \strong{cumulative daily weather data}, a \CRANpkg{data.table} containing:
\itemize{
\item \strong{cdd}, cumulative degree days,
\item \strong{cwh}, cumulative wet hours,
\item \strong{cr}, cumulative rainfall in mm,
\item \strong{gp_standard}, standard growing points assuming growth is not impeded
by infection,
}
\item \strong{infected_coords}, a \CRANpkg{data.table} of only infectious growing
point coordinates,
\item \strong{new_infections}, a \CRANpkg{data.table} of newly infected growing
points,
\item \strong{exposed_gp}, a \CRANpkg{data.table} of exposed growing points in the
latent period phase of infection.
}
}
\description{
Simulate the spatiotemporal development of Ascochyta blight in a chickpea
paddock over a growing season on a one square metre basis.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# First weather data needs to be imported and formatted with `format_weather`
Newmarracarra <-
   read.csv(system.file("extdata",
            "1998_Newmarracarra_weather_table.csv", package = "ascotraceR"))
station_data <-
   system.file("extdata", "stat_dat.csv", package = "ascotraceR")

weather_dat <- format_weather(
   x = Newmarracarra,
   POSIXct_time = "Local.Time",
   temp = "mean_daily_temp",
   ws = "ws",
   wd_sd = "wd_sd",
   rain = "rain_mm",
   wd = "wd",
   station = "Location",
   time_zone = "Australia/Perth",
   lonlat_file = station_data)

# Now the `trace_asco` function can be run to simulate disease spread
traced <- trace_asco(
  weather = weather_dat,
  paddock_length = 100,
  paddock_width = 100,
  initial_infection = "1998-06-10",
  sowing_date = as.POSIXct("1998-06-09"),
  harvest_date = as.POSIXct("1998-06-09") + lubridate::ddays(70), # run the model for 70 days
  time_zone = "Australia/Perth",
  gp_rr = 0.0065,
  primary_inoculum_intensity = 1000,
  spores_per_gp_per_wet_hour = 0.22,
  primary_infection_foci = "centre")

traced[[70]] # extracts the model output for day 70
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=tidy_trace]{tidy_trace()}}, \code{\link[=summarise_trace]{summarise_trace()}}
}
