% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace_asco.R
\name{trace_asco}
\alias{trace_asco}
\title{Simulates the spread of Ascochyta blight in a chickpea field}
\usage{
trace_asco(
  weather,
  paddock_length,
  paddock_width,
  sowing_date,
  harvest_date,
  initial_infection,
  seeding_rate = 40,
  gp_rr = 0.0065,
  max_gp_lim = 15000,
  max_new_gp = 350,
  latent_period_cdd = 200,
  time_zone = "UTC",
  primary_infection_foci = "random",
  primary_infection_intensity = 1,
  n_foci = 1,
  spores_per_gp_per_wet_hour = 0.22
)
}
\arguments{
\item{weather}{weather data, recorded by a local weather station, over a chickpea
growing season for the model operation}

\item{paddock_length}{length of a paddock in metres (y)}

\item{paddock_width}{width of a paddock in metres (x)}

\item{sowing_date}{a character string of a date value indicating sowing
date of chickpea seed and the start of the ascotraceR model. Preferably
in \acronym{ISO8601} format (YYYY-MM-DD), \emph{e.g.} \dQuote{2020-04-26}.
Assumes there is sufficient soil moisture to induce germination and start the
crop growing season.}

\item{harvest_date}{a character string of a date value indicating harvest date of
chickpea crop, which is also the last day to run the ascotraceR model. Preferably in
\acronym{ISO8601} format (YYYY-MM-DD), \emph{e.g.} \dQuote{2020-04-26}.}

\item{initial_infection}{refers to initial or primary infection on seedlings,
resulting in the production of infected growing points}

\item{seeding_rate}{indicate the rate at which chickpea seed is sown per
square metre. Defaults to \code{40}}

\item{gp_rr}{refers to rate of increase in chickpea growing points
per degree Celsius per day. Defaults to \code{0.0065}}

\item{max_gp_lim}{Maximum number of chickpea growing points (meristems) allowed
per square meter. Defaults to \code{15000}.}

\item{max_new_gp}{Maximum number of new chickpea growing points (meristems)
which develop per day, per square meter. Defaults to \code{350}.}

\item{latent_period_cdd}{latent period in cumulative degree days (sum of
daily temperature means) is the period between infection and production of
lesions on susceptible growing points. Defaults to \code{200}}

\item{time_zone}{refers to time in Coordinated Universal Time (UTC)}

\item{primary_infection_foci}{it refers to the inoculated quadrat
located at the centre of the paddock from where disease spreads
Defaults to \code{"centre"}}

\item{primary_infection_intensity}{The intensity of the starting epidemic as
described by the number of number of sporulating growing points.}

\item{n_foci}{only relevant when primary_infection_foci = "random" and notes the number
of primary_infection_foci at initial infection.}

\item{spores_per_gp_per_wet_hour}{Number of spores produced per sporulating growing point each wet hour.
Also known as the 'spore_rate'. Value is dependent on the susceptibility of the host genotype.}
}
\value{
a x y \code{data.frame} simulating the spread of Ascochyta blight in a
chickpea paddock
}
\description{
\code{trace_asco} simulates the spatiotemporal development of Ascochyta blight in a chickpea field
over a growing season
}
\examples{
First weather data needs to be imported and formatted with `format_weather`
Billa_Billa <-
  readr::read_csv(system.file("extdata",
                       "2020_Billa_Billa_weather.csv", package = "ascotraceR")) \%>\%
   # convert wind direction recorded as text to degrees
  dplyr::mutate(
    wind_direction = dplyr::case_when(
      max_wind_direction == "N" ~ "0",
      max_wind_direction == "NbE" ~ "11.25",
      max_wind_direction == "NNE" ~ "22.5",
      max_wind_direction == "NEbN" ~ "33.75",
      max_wind_direction == "NE" ~ "45",
      max_wind_direction == "NEbE" ~ "56.25",
      max_wind_direction == "ENE" ~ "67.5",
      max_wind_direction == "EbN" ~ "73.5",
      max_wind_direction == "E" ~ "90",
      max_wind_direction == "EbS" ~ "101.2",
      max_wind_direction == "ESE" ~ "112.5",
      max_wind_direction == "SEbE" ~ "123.8",
      max_wind_direction == "SE" ~ "135.1",
      max_wind_direction == "SEbS" ~ "146.3",
      max_wind_direction == "SSE" ~ "157.6",
      max_wind_direction == "SbE" ~ "168.8",
      max_wind_direction == "S" ~ "180",
      max_wind_direction == "SbW" ~ "191.2",
      max_wind_direction == "SSW" ~ "202.5",
      max_wind_direction == "SWbS" ~ "213.8",
      max_wind_direction == "SW" ~ "225",
      max_wind_direction == "SWbW" ~ "236.2",
      max_wind_direction == "WSW" ~ "247.5",
      max_wind_direction == "WbS" ~ "258.8",
      max_wind_direction == "W" ~ "270",
      max_wind_direction == "WbN" ~ "281.2",
      max_wind_direction == "WNW" ~ "292.5",
      max_wind_direction == "NWbW" ~ "303.8",
      max_wind_direction == "NW" ~ "315",
      max_wind_direction == "NWbN" ~ "326.2",
      max_wind_direction == "NNW" ~ "337.5",
      max_wind_direction == "NbW" ~ "348.8",
      TRUE ~ max_wind_direction
    )
  ) \%>\%
  dplyr::mutate(wind_direction = as.numeric(wind_direction)) \%>\%
  dplyr::rename(wd = wind_direction) \%>\%
  dplyr::rename(stationID=station) \%>\%
  dplyr::rename(ws =  `avg_wind_speed (km/h)`) \%>\%
  dplyr::mutate(wd_sd = as.numeric(sd(wd))) \%>\%
  dplyr::mutate(ws_sd = as.numeric(sd(ws)))

station_data <-
  system.file("extdata", "stat_dat.csv", package = "ascotraceR")

weather_dat <- format_weather(x = Billa_Billa,
  POSIXct_time = "local_time",
  temp = "mean_daily_temperature",
  ws = "ws",
  wd_sd = "wd_sd",
  rain = "rain",
  wd = "wd",
  station = "location",
  time_zone = "Australia/Brisbane",
  lonlat_file = station_data)



traced <- trace_asco(
  weather = weather_dat,
  paddock_length = 20,
  paddock_width = 20,
  initial_infection = "2020-07-16",
  sowing_date = as.POSIXct("2020-06-04"),
  harvest_date = as.POSIXct("2020-06-04") + lubridate::ddays(30),
  time_zone = "Australia/Brisbane",
  primary_infection_foci = "center")
  traced[[5]]
}
