% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_paddock_gp.R
\name{update_paddock_gp}
\alias{update_paddock_gp}
\title{Update chickpea growing points over a whole paddock}
\usage{
update_paddock_gp(paddock_coords, mean_air_temp, gp_rr, max_gp)
}
\arguments{
\item{paddock_coords}{a data.frame of paddock coordinates containing a column variable "noninfected_gp}

\item{mean_air_temp}{mean daily air temperature}

\item{gp_rr}{growing points replication rate}

\item{max_gp}{maximum growing points}
}
\value{
a vector of updated growing points at each paddock location
}
\description{
Update chickpea growing points over a whole paddock
}
\examples{
paddock <- expand.grid(x = 1:40,
                       y = 1:40)
paddock$noninfected_gp <- sample(c(-1,0), nrow(paddock), replace = TRUE)

GPs <- update_paddock_gp(paddock_coords = paddock,
                         mean_air_temp = 24,
                         gp_rr = 0.0065,
                         max_gp = 15000
                         )
GPs
}
